+++ auto_route/lib/auto_route.dart
@@ -1,0 +1,6 @@
+export 'src/router_utils.dart';
+
+export 'src/extended_navigator.dart';
+export 'src/route_guard.dart';
+export 'src/auto_route_wrapper.dart';
+export 'src/transitions_builders.dart';