+++ auto_route_generator/lib/route_config_visitor.dart
@@ -1,5 +1,6 @@
 export 'src/custom_transtion_builder.dart';
 export 'src/route_config.dart';
 export 'src/router_config.dart';
-export 'src/route_parameter.dart';
+export 'src/route_paramter_config.dart';
+export 'src/route_gaurd_config.dart';
 export 'src/router_config_visitor.dart';