+++ example/lib/model.dart
@@ -1,3 +1,0 @@
-class Model {
-  String value = "model value";
-}