+++ example/lib/router.dart
@@ -1,16 +1,39 @@
+import 'dart:async';
+
+import 'package:auto_route/auto_route.dart';
 import 'package:auto_route/auto_route_annotations.dart';
 import 'package:example/screens/home_screen.dart';
 import 'package:example/screens/login_screen.dart';
 import 'package:example/screens/second_screen.dart';
+import 'package:example/screens/third_screen.dart';
+import 'package:flutter/material.dart';
 
-@AutoRouter(generateNavigator: true, generateRouteList: false)
+@CustomAutoRouter(
+    transitionsBuilder: TransitionsBuilders.slideLeft,
+    opaque: true,
+    barrierDismissible: true,
+    durationInMilliseconds: 300,
+    generateRouteList: true)
+// @CupertinoAutoRouter()
 class $Router {
-  @MaterialRoute(initial: true)
-  HomeScreen homeScreenRoute;
+  @initial
+  HomeScreen homeScreen;
 
-  @CupertinoRoute(name: '/custom_cupertino_name')
-  SecondScreen secondScreenRoute;
+  @GuardedBy([AuthGuard])
+  SecondScreen secondScreen;
 
-  @CustomRoute(name: '/custom_route_name')
+  @GuardedBy([AuthGuard])
+  ProfileScreen profileScreen;
+
+  @CustomRoute(returnType: bool)
   LoginScreen loginScreenDialog;
 }
+
+class AuthGuard extends RouteGuard {
+  @override
+  Future<bool> canNavigate(
+      BuildContext context, String routeName, Object arguments) async {
+     return true;
+  }
+}
+