+++ auto_route/CHANGELOG.md
@@ -1,3 +1,11 @@
+## [0.3.0] Breaking Changes!
+- Add global route customization
+Use MaterialAutoRouter, CupertinoAutoRouter or CustomAutoRouter instead of AutoRouter
+- Navigator key is now accessed by calling Router.navigator.key instead of Router.navigatorKey.
+- Add Route guards
+- Add optional returnType param in all AutoRoute types.
+- Remove generate Navigator Key optionality.
+- Fix generate error when using dynamic vars in route widget constructors
 ## [0.2.2]
 
 - Add option to generate a list with all route names